structured the code
