+++ src/main.py
@@ -14,57 +14,7 @@ from train import *
 from model import *
 
 if __name__ == '__main__':
-    print('h')
-    '''
-    opts = [NoamOpt(512, 1, 4000, None), 
-            NoamOpt(512, 1, 8000, None),
-            NoamOpt(256, 1, 4000, None)]
-    plt.plot(np.arange(1, 20000), [[opt.rate(i) for opt in opts] for i in range(1, 20000)])
-    plt.legend(["512:4000", "512:8000", "256:4000"])
-    plt.show()
-    
-    # Example of label smoothing.
-    crit = LabelSmoothing(5, 0, 0.4)
-    predict = torch.FloatTensor([[0, 0.2, 0.7, 0.1, 0],
-                                 [0, 0.2, 0.7, 0.1, 0], 
-                                 [0, 0.2, 0.7, 0.1, 0]])
-    v = crit(Variable(predict.log()), 
-             Variable(torch.LongTensor([2, 1, 0])))
-    
-    # Show the target distributions expected by the system.
-    plt.imshow(crit.true_dist)
-    plt.show()
-    
-    crit = LabelSmoothing(5, 0, 0.1)
-    def loss(x):
-        d = x + 3 * 1
-        predict = torch.FloatTensor([[0, x / d, 1 / d, 1 / d, 1 / d],])
-        v = Variable(predict.log())
-        vv = Variable(torch.LongTensor([1]))
-        return crit(v, vv).data
-    plt.plot(np.arange(1, 100), [loss(x) for x in range(1, 100)])
-    plt.show()
-    
-    # Train the simple copy task.
-    V = 11
-    criterion = LabelSmoothing(size=V, padding_idx=0, smoothing=0.0)
-    model = make_model(V, V, N=2)
-    model_opt = NoamOpt(model.src_embed[0].d_model, 1, 400,
-            torch.optim.Adam(model.parameters(), lr=0, betas=(0.9, 0.98), eps=1e-9))
-    
-    for epoch in range(10):
-        model.train()
-        run_epoch(data_gen(V, 30, 20), model, 
-                  SimpleLossCompute(model.generator, criterion, model_opt))
-        model.eval()
-        print(run_epoch(data_gen(V, 30, 5), model, 
-                        SimpleLossCompute(model.generator, criterion, None)))
-    
-    model.eval()
-    src = Variable(torch.LongTensor([[1,2,3,4,5,6,7,8,9,10]]) )
-    src_mask = Variable(torch.ones(1, 1, 10) )
-    print(greedy_decode(model, src, src_mask, max_len=10, start_symbol=1))
-    '''
+
     import spacy
     spacy_fr = spacy.load('fr_core_news_sm')
     spacy_en = spacy.load('en_core_web_sm')
